Extra Setup: Checkpatch on PATH and Mutt keybinds
Added additional configuration options for adding `checkpatch.pl` to a
folder already on the PATH and added notes for including the keybinds
I found useful in *mutt* when doing the peer review portion of the
assignments.

The mutt configuration could be moved to other documents that may be
more relevant, but people are more likely to read the setup assignment.